[ACS-80] Restrict source address of clients accessing API methods.
It was reported that we are not currently using the RBAC model because we need to control the root admin access in public interfaces. RBAC does not provide means to restrict users giving their source IP. This control is achieved via “commands.properties” in ACS. Management servers, which are dedicated to public access have API methods (admin ones) restricted (via the "commands.properties" file). This means that, if the root admin access Cloud Stack via a public IP, he/she will not be able to execute root methods. However, he/she would still be able to view information. We then use an “internal” only management server that has the “commands.properties” configured to give permissions to root admins.

The problem is that the use of “commands.properties” restricts us to use only the default roles. Whereas, with RBAC, one can customize roles as she/he desires. This means, it (with RBAC) is possible to create roles that provide “view only” access to the system. Other customizations are also possible to address the most different business requirements. However, the lack of means to restrict the source IP of clients is a real security problem for us.

We Decided to implement Wido’s approach. If we ever find the need to something else, we can always extend it in the future.