[ACS-47] View volumes from primary storage not working when storage UUID is not a UUID
While doing some tests I found the following problem regarding volumes listing of primary storage.

When configuration a pre-setup primary storage we can enter the name-label of the storage that is going to be used by ACS and is already set up in the host. The problem is that we can use any String of characters there, and this String does not need to be a UUID one. When listing volumes from a primary storage that has such conditions, the list will return all of the volumes in the cloud because the ad-hoc “API framework” will ignore that value as it is not a UUID type.

The issue is being tracked in CloudStack's Jira via https://issues.apache.org/jira/browse/CLOUDSTACK-10276.